made initialisations through array, set one listener to all number buttons
